adjustments in the controller and helper file
